move under Symfony\FrameworkBundle, as dependent on it
