+++ tests/Rector/Contrib/Symfony/FrameworkBundle/GetParameterToConstructorInjectionRector/Correct/correct2.php.inc
