+++ tests/Rector/Contrib/Symfony/FrameworkBundle/GetParameterToConstructorInjectionRector/config.yml
@@ -1,0 +1,2 @@
+rectors:
+    Rector\Rector\Contrib\Symfony\FrameworkBundle\GetParameterToConstructorInjectionRector: ~
